Fix driver stuck on RecursionError on COMMIT SUCCESS (#1192)
The driver could get stuck when the result summary was so deeply nested that it
caused an `RecursionError`. It would also get stuck on certain protocol
violations in the same place (e.g., unknown PackStream tag). However, the
latter is much more unlikely to occur.

This patch will make the driver ditch the connection when it's not able to
properly read a message for unexpected reasons and let the error bubble up.

* Add unit tests